AspNetCore OAuth sample
This provides an OAuth sample implementation for Asp.Net Core.

This is my first attempt at writing my own OAuth provider for Asp.Net
Core and I'm also fairly new to OAuth in general. So please let me know
if there is anything that can be improved in this sample

closes #45