Use VssAzureIdentityCredential in examples
This has a number of advantages over the approaches shown here. We get
things like automatic token refreshes for free, and the logic is
overall much clearer. I threw in some drive-by package updates to
Azure.Identity dependencies, because `dotnet build` was complaining that
the versions referenced had security issues.